[STRY/BE] REST API: Bullets (#27)
* Added skeleton of bullet.js based on Ryan's demo

* Read in data from the POST request

* Added database to gitignore

* Finished basic read of a bullet from id

* POST request to create bullet is functional

* GET request to get bullet from ID is functional

* Improved comments to be slightly more descriptive

* Removed console output from \read\bullet

* Bullet create and read now send  error response

* Created bullet delete

* Added thrown error on missing field and success response

* Removing unnecessary console.log on field missing error

* Changed result to response for consistency

All three other CRUD routes use response instead of result, and this
aligns create with them.

* Created update

Update will take a json field of the form ID and updateField, with
updateField containing any fields the request wishes to update.
It will take that input and first duplicate the original response, then
add in the new updated fields from updateField. It will then return
a success response.

* Updated comment to reflect error behavior

A missing field always throws an error in the code as written, so the
comment should reflect that.

* Added database response to creating bullet

Previously, with randomly generated IDs, there would be no way to
request the data afterwards, as the ID would be unknown and lost forever.
Now, the full database response will be sent for frontend use.

* Change database name to db

* Adjusted fields and added comments for them

Added a docType that keeps track of exactly what kind of document it is
and explained what every field of the bullets is intended for

* Added landing page html and uploaded logo

* Create node.js.yml

* Temporarily comment out build job

* Update workflow with temporary starter job

* added login page html

* Replaced sample job with lint job

* Installed and configured eslint

* Installed google config for eslint

* Added exception to linebreak style rule

* Formatted start.js file to abide by style guide

* Removed old db name from gitignore

* Removed docType from bullet creation

It is always the same on bullet creation, so passing it is unnecessary.

* Installed jest for unit tests

* Added diagram for phase 1 of CI pipeline

* Update source/routes/update/bullet.js

Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>

* Apply suggestions from automatic style guideline suggestions

Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>

* Reimplment lint job that got overwritten (#41)

* Updated bullet CRUD to implement daily bullet lists

* Add files via upload

* Delete daily.js

* Move pipeline diagram to appropirate folder

* Fixed response overriding

The response of the initial document get request and the parent get
request were overriding one another, causing the bullets array to
become null.

Co-authored-by: Zhongqian <zhl024@ucsd.edu>

* Upload markdown of status

* Optimized the update function

* Apply suggestions from automatic code review

Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>

* Added skeleton of bullet.js based on Ryan's demo

* Read in data from the POST request

* Added database to gitignore

* Finished basic read of a bullet from id

* POST request to create bullet is functional

* GET request to get bullet from ID is functional

* Improved comments to be slightly more descriptive

* Removed console output from \read\bullet

* Bullet create and read now send  error response

* Created bullet delete

* Added thrown error on missing field and success response

* Removing unnecessary console.log on field missing error

* Changed result to response for consistency

All three other CRUD routes use response instead of result, and this
aligns create with them.

* Created update

Update will take a json field of the form ID and updateField, with
updateField containing any fields the request wishes to update.
It will take that input and first duplicate the original response, then
add in the new updated fields from updateField. It will then return
a success response.

* Updated comment to reflect error behavior

A missing field always throws an error in the code as written, so the
comment should reflect that.

* Added database response to creating bullet

Previously, with randomly generated IDs, there would be no way to
request the data afterwards, as the ID would be unknown and lost forever.
Now, the full database response will be sent for frontend use.

* Change database name to db

* Adjusted fields and added comments for them

Added a docType that keeps track of exactly what kind of document it is
and explained what every field of the bullets is intended for

* Removed old db name from gitignore

* Removed docType from bullet creation

It is always the same on bullet creation, so passing it is unnecessary.

* Updated bullet CRUD to implement daily bullet lists

* Update source/routes/update/bullet.js

Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>

* Apply suggestions from automatic style guideline suggestions

Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>

* Fixed response overriding

The response of the initial document get request and the parent get
request were overriding one another, causing the bullets array to
become null.

Co-authored-by: Zhongqian <zhl024@ucsd.edu>

* Add files via upload

* Delete daily.js

* Optimized the update function

* Applied eslint --fix to conform to the style guide

* Added authentification middleware

* Added comments detailing the JSON request body

Also updated some id to _id to reflect the proper JSON request

* Removed middleware authentification

* Fixed potential hang

If putting itself in the daily document bullet list errored out, it
would not return an error and would instead pretend to be
working. Tested to be working, now.

Co-authored-by: Lars Vlahakis <lvlahaki@ucsd.edu>
Co-authored-by: Brandon Chaing <60688306+bchaing@users.noreply.github.com>
Co-authored-by: Ivy <46914183+ivychxn@users.noreply.github.com>
Co-authored-by: Brandon Chaing <bchaing@ucsd.edu>
Co-authored-by: b2wu <81795401+b2wu@users.noreply.github.com>
Co-authored-by: Zhongqian Li <64935761+zhl024@users.noreply.github.com>
Co-authored-by: Zhongqian <zhl024@ucsd.edu>